refactor: 예약 삭제시 로직 수정

+++ src/main/java/roomescape/payment/PaymentService.java
@@ -1,7 +1,5 @@
 package roomescape.payment;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import roomescape.controller.reservation.dto.CreateReservationRequest;
@@ -11,12 +9,12 @@ import roomescape.payment.dto.CancelPaymentRequest;
 import roomescape.payment.dto.CreatePaymentRequest;
 import roomescape.payment.dto.PaymentConfirmResponse;
 import roomescape.repository.PaymentRepository;
-import roomescape.service.exception.PaymentInfoNotFoundException;
+
+import java.util.Optional;
 
 @Service
 public class PaymentService {
 
-    private static final Logger log = LoggerFactory.getLogger(PaymentService.class);
     private final PaymentClient paymentClient;
     private final PaymentRepository paymentRepository;
 
@@ -35,12 +33,12 @@ public class PaymentService {
 
     @Transactional
     public void deletePayment(final long reservationId) {
-        try {
-            final Payment payment = paymentRepository.fetchByReservationId(reservationId);
-            paymentClient.cancelPayment(new CancelPaymentRequest(payment.getPaymentKey(), "단순변심"));
-            paymentRepository.delete(payment);
-        } catch (final PaymentInfoNotFoundException ignore) {
-            log.warn("결제 정보가 존제하지 않습니다.");
+        final Optional<Payment> findPayment = paymentRepository.findByReservationId(reservationId);
+        if (findPayment.isEmpty()) {
+            return;
         }
+        final Payment payment = findPayment.get();
+        paymentClient.cancelPayment(new CancelPaymentRequest(payment.getPaymentKey(), "단순변심"));
+        paymentRepository.delete(payment);
     }
 }
